backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-multipart
python-jose[cryptography]
email-validator
pillow
pytesseract
pymupdf

